Static Sidebar implemented
Changes to be committed:
	modified:   client/src/App.css
	modified:   client/src/components/builder/index.js
	new file:   client/src/components/builder/indexOld.js
	modified:   client/src/components/builder/rawJson.js
	modified:   client/src/components/viewer/cardViews.js
	modified:   client/src/components/viewer/index.js
	modified:   client/src/layouts/MainNavBar.js
	modified:   client/src/layouts/ObjectNavBar.js
	modified:   client/src/layouts/ObjectViewLayout.js
	new file:   client/src/layouts/objectSideBar.js
	modified:   client/src/routes.js
