+++ client/src/App.css
@@ -73,29 +73,22 @@
 /* Outlet styles */
 .outlet-root {
   display: flex;
-  height: 100%;
+  height: 100vh;
   overflow: hidden;
   width: 100%;
-}
-
-.outlet-wrapper {
-  display: flex;
-  flex: 1 1 auto;
-  overflow: hidden;
-  padding-top: 64;
+  flex-direction: column;
 }
 
 .outlet-container {
   display: flex;
   flex: 1 1 auto;
   overflow: hidden;
+  padding-top: 75px;
 }
 
 .outlet-content {
   flex: 1 1 auto;
-  height: 100%;
-  overflow: auto;
-  padding-top: 75px;
+  overflow: auto; /* change back to auto ? */
 }
 
 /* NavBar styles */
@@ -121,3 +114,19 @@
     display: none;
   }
 }
+
+/* Object Styles */
+.object-sidebar {
+  width: 200px;
+  flex-shrink: 0;
+  overflow: auto;
+  height: 100%;
+}
+
+.object-contents {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  overflow: auto;
+}
+
