+++ client/src/components/builder/rawJson.js
@@ -2,6 +2,7 @@ import { Box, Button, Card, CardContent, CardHeader, TextField } from "@material
 import React, { useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { updateBco, updateModified } from "../../slices/bcoSlice";
+
 export const RawJson = () => {
   const dispatch = useDispatch()
   const [bco, setBco] = useState(useSelector(state => state.bco.data))
