+++ client/src/components/viewer/index.js
@@ -1,131 +1,28 @@
 // src/components/viewer/index.js
 
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import {
-  Button,
   Card,
-  CardActions,
-  CardHeader,
   CardContent,
-  Collapse,
   Container,
   Grid,
-  ListItem,
-  ListItemText,
   Typography
 } from "@material-ui/core";
-import { useNavigate } from "react-router-dom";
-import DataObjectIcon from "@mui/icons-material/DataObject";
 import { useDispatch, useSelector } from "react-redux";
 import PropTypes from "prop-types";
 import {
   ProvenanceView, UsabilityView, DescriptionView, ExtensionView,
-  ExecutionView, ParametricView, IoView, ErrorView
+  ExecutionView, ParametricView, IoView, ErrorView, RawJson, TreeView
 } from "./cardViews";
-import { handleDownloadClick } from "../fileHandeling";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import { getPubBco, deriveBco } from "../../slices/bcoSlice";
-import { styled } from "@mui/material/styles";
-import IconButton from "@mui/material/IconButton";
-
-const data = [
-  {
-    name: "Provenance Domain",
-    value: "provenance_domain"
-  },
-  {
-    name: "Usability Domain",
-    value: "usability_domain"
-  },
-  {
-    name: "Description Domain",
-    value: "description_domain"
-  },
-  {
-    name: "Extension Domain (Optional)",
-    value: "extension_domain"
-  },
-  {
-    name: "Parametric Domain",
-    value: "parametric_domain"
-  },
-  {
-    name: "IO Domain",
-    value: "io_domain"
-  },
-  {
-    name: "Execution Domain",
-    value: "execution_domain"
-  },
-  {
-    name: "Error Domain",
-    value: "error_domain"
-  }
-];
-
-const ExpandMore = styled((props) => {
-  const { expand, ...other } = props;
-  return <IconButton {...other} />;
-})(({ theme, expand }) => ({
-  transform: !expand ? "rotate(0deg)" : "rotate(180deg)",
-  marginLeft: "auto",
-  transition: theme.transitions.create("transform", {
-    duration: theme.transitions.duration.shortest,
-  }),
-}));
+import { getPubBco } from "../../slices/bcoSlice";
+import "../../App.css"
+import { useOutletContext } from "react-router-dom";
 
 export default function BcoViewer () {
-  const navigate = useNavigate();
+  const {domain, setDomain} = useOutletContext()
   const dispatch = useDispatch();
-  const [value, setValue] = React.useState(0);
   const bco = useSelector(state => state.bco.data)
-  const [expanded, setExpanded] = useState(false);
-  const jsonData = useSelector((state) => state.bco.data);
-
-  const handleExpandClick = () => {
-    setExpanded(!expanded);
-  };
-
-  const handleChange = (newValue) => {
-    setValue(newValue);
-  };
-
-  const  handleDerive = (jsonData) => {
-    navigate("/builder")
-    dispatch(deriveBco(jsonData))
-  };
-
-  function a11yProps(index) {
-    return {
-      id: `simple-tab-${index}`,
-      "aria-controls": `simple-tabpanel-${index}`,
-    };
-  }
-
-  function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-    return (
-      <div
-        role="tabpanel"
-        hidden={value !== index}
-        id={`simple-tabpanel-${index}`}
-        aria-labelledby={`simple-tab-${index}`}
-        {...other}
-      >
-        {value === index && (
-          <Container>
-            {children}
-          </Container>
-        )}
-      </div>
-    );
-  }
-  TabPanel.propTypes = {
-    children: PropTypes.node,
-    index: PropTypes.any.isRequired,
-    value: PropTypes.any.isRequired,
-  };
-
+  console.log(domain)
   function validURL(url) {
     try {
       new URL(url);
@@ -149,53 +46,34 @@ export default function BcoViewer () {
         });
     }
   }, [])
-  return (
-    
-    <Container>
+
+  function TabPanel(props) {
+    const { children, domain, index, ...other } = props;
+    return (
+      <div
+        role="tabpanel"
+        hidden={domain !== index}
+        id={`simple-tabpanel-${index}`}
+        aria-labelledby={`simple-tab-${index}`}
+        {...other}
+      >
+        {domain === index && (
+          <Container>
+            {children}
+          </Container>
+        )}
+      </div>
+    );
+  }
+  TabPanel.propTypes = {
+    children: PropTypes.node,
+    index: PropTypes.any.isRequired,
+    domain: PropTypes.any.isRequired,
+  };
+
+  return (  
+    <>
       <Grid container spacing={2}>
-        <Card className='sidebar'>
-          <Grid item>
-            {data.map((item, index) => (
-              <ListItem selected={ value===index } value={index} button key={index} onClick={() => {handleChange(index)}} {...a11yProps(index)}>
-                <DataObjectIcon />{" "}<ListItemText primary={item.name} />
-              </ListItem>
-            ))}
-            <br/>
-            <CardActions disableSpacing>
-              <CardHeader title="Tools"/>
-              <ExpandMore
-                expand={expanded}
-                onClick={handleExpandClick}
-                aria-expanded={expanded}
-                aria-label="show more"
-              >
-                <ExpandMoreIcon />
-              </ExpandMore>
-            </CardActions>
-            <Collapse in={expanded} timeout="auto" unmountOnExit>
-              <br/>
-              <Grid item>
-                <Button 
-                  className="download-button"
-                  type='submit'
-                  variant="contained"
-                  color="primary"
-                  onClick={() => {handleDownloadClick(jsonData)}}
-                > Download BCO</Button>
-              </Grid>
-              <br/>
-              <Grid item>
-                <Button 
-                  className="derive-button"
-                  type='submit'
-                  variant="contained"
-                  color="primary"
-                  onClick={() => {handleDerive(jsonData)}}
-                > Derive BCO </Button>
-              </Grid>
-            </Collapse>
-          </Grid>
-        </Card>
         <Grid item xs={12} md>
           <Card>
             <CardContent>
@@ -211,33 +89,38 @@ export default function BcoViewer () {
             </CardContent>
           </Card>
           <br/>
-          <TabPanel value={value} index={0}>
+          <TabPanel domain={domain} index={0}>
             <ProvenanceView/>
           </TabPanel>
-          <TabPanel value={value} index={1}>
+          <TabPanel domain={domain} index={1}>
             <UsabilityView/>
           </TabPanel>
-          <TabPanel value={value} index={2}>
+          <TabPanel domain={domain} index={2}>
             <DescriptionView/>
           </TabPanel>
-          <TabPanel value={value} index={3}>
+          <TabPanel domain={domain} index={3}>
             <ExtensionView/>
           </TabPanel>
-          <TabPanel value={value} index={4}>
+          <TabPanel domain={domain} index={4}>
             <ParametricView/>
           </TabPanel>
-          <TabPanel value={value} index={5}>
+          <TabPanel domain={domain} index={5}>
             <IoView/>
           </TabPanel>
-          <TabPanel value={value} index={6}>
+          <TabPanel domain={domain} index={6}>
             <ExecutionView/>
           </TabPanel>
-          <TabPanel value={value} index={7}>
+          <TabPanel domain={domain} index={7}>
             <ErrorView/>
+          </TabPanel>
+          <TabPanel domain={domain} index={8}>
+            <RawJson/>
+          </TabPanel>
+          <TabPanel domain={domain} index={9}>
+            <TreeView/>
           </TabPanel>
         </Grid>
       </Grid>
-    </Container>
-
+    </>
   )
 }
