+++ client/src/layouts/ObjectViewLayout.js
@@ -2,17 +2,23 @@
 
 import React from "react";
 import { Outlet } from "react-router-dom";
-import NavBar from "./ObjectNavBar";
+import ObjectNavBar from "./ObjectNavBar";
+import ObjectSideBar from "./objectSideBar";
 import "../App.css";
 
 export default function ObjectLayout ()  {
-
+  const [domain, setDomain] = React.useState(0);
   return (
     <div className="outlet-root">
+      <ObjectNavBar />
       <div className="outlet-container">
+        <ObjectSideBar
+          className="object-sidebar"
+          domain={domain}
+          setDomain={setDomain}
+        />
         <div className="outlet-content">
-          <NavBar />
-          <Outlet />
+          <Outlet context={{domain, setDomain}}/>
         </div>
       </div>
     </div>
